requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Login==0.6.3
mysql-connector-python==8.3.0
razorpay==1.4.1
itsdangerous==2.1.2
Werkzeug==2.3.7
python-dotenv==1.0.1